refactor(LLC): reuse getAccountTuplesForCurrency
