+++ .changeset/eight-parrots-cry.md
@@ -1,0 +1,6 @@
+---
+"ledger-live-desktop": minor
+"@ledgerhq/live-common": minor
+---
+
+move to reuse getAccountTuplesForCurrency
